test 4

+++ database/migrations/2019_10_30_170802_create_table_client.php
@@ -19,7 +19,7 @@ class CreateTableClient extends Migration
             $table->string('prenom');
             $table->string('nationalite');
             $table->string('typePiece');
-            $table->integer('numeroPiece')->unique();
+            $table->bigInteger('numeroPiece')->unique();
             $table->string('adresse');
             $table->string('telephone')->unique();
             $table->string('email')->unique();
